+++ src/speconv/core.clj
@@ -35,14 +35,16 @@
 
 (defmulti export
   "Take a specified input map and produced some output.  The output is
-  not validated against any spec, but the input is."
-  (fn [to entity]
-    to))
+  not validated against any spec, but the input is.  The context
+  argument provides some description as to the destination the data is
+  being exported to and should typically be a keyword, such as :api."
+  (fn [from to-context entity]
+    [from to-context]))
 
 (defmacro exporter
-  [from [entity-sym] & body]
-  `(defmethod export ~from
-     [_# ~entity-sym]
+  [from to-context [entity-sym] & body]
+  `(defmethod export [~from ~to-context]
+     [_# _# ~entity-sym]
      (if (s/valid? ~from ~entity-sym)
        (do ~@body)
        (throw (ex-info (str "Input is not a valid " ~from)
@@ -50,14 +52,16 @@
 
 (defmulti import
   "Take some input and produce a specified output.  The input is not
-  validated against any spec, but the output is."
-  (fn [from entity]
-    from))
+  validated against any spec, but the output is.  The context argument
+  provides some description of the source of the data and should
+  typically be a keyword, such as :config."
+  (fn [to from-context entity]
+    [to from-context]))
 
 (defmacro importer
-  [to [entity-sym] & body]
-  `(defmethod import ~to
-     [_# ~entity-sym]
+  [to from-context [entity-sym] & body]
+  `(defmethod import [~to ~from-context]
+     [_# _# ~entity-sym]
      (let [result# (do ~@body)]
        (if (s/valid? ~to result#)
          result#